Check number of arguments for commands
Uses valid usages to verify number of arguments is applicable. This means it should be assured that all usages for commands are properly documented. Note that this allows extraneous arguments, to allow for commands like anon that take the entire section after the command. For now this is fine because extraneous arguments will be ignored anyways.