Allow commands with separator in them
This was actually kind of complicated to implement. The main reason to add this is to support composite commands. A warning is that command aliases should never start with the same thing, eg. there should not be a command ">xp" as well as ">xp something", because there is no guarantee for which it will match first.