Fix commands and send to whitelisted channels
Fixes bug introduced by 9771035d2ecd1c3dc7018adcf4c553b266fbbd03. Because the same command could be stored under multiple aliases in the map, this was producing duplicates in the commands output.

Merge sending command related messages to a class and change the behavior: instead of always deleting the message and sending to DMs if in a guild, we can just output commands directly if it is in a whitelisted channel (because these channels are already fine to be filled with output).