Generic moderation roles and quarantine
This is not the best thing but it does allow for a limited scope of generic moderation roles. Right now it only supports one moderation role at a time. In the future we may want to support multiple. Ideally, we would also have a role managing system instead of doing it in the moderation module.